# backend/spec/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/workspaces/{workspaceId}/goals": {
            "get": {
                "operationId": "GoalsController_list",
                "parameters": [
                    {
                        "name": "include_archived",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "default": "false",
                            "type": "string",
                            "enum": ["true", "false"]
                        }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "workspace_id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "target_amount": {
                                                "type": "number"
                                            },
                                            "initial_amount": {
                                                "type": "number"
                                            },
                                            "currency": {
                                                "type": "string"
                                            },
                                            "target_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "plan_start_month": {
                                                "type": "string"
                                            },
                                            "archived_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            },
                                            "current_amount": {
                                                "type": "number"
                                            },
                                            "remaining_amount": {
                                                "type": "number"
                                            },
                                            "progress_percentage": {
                                                "type": "number"
                                            },
                                            "remaining_months": {
                                                "type": "integer"
                                            },
                                            "recommended_monthly_amount": {
                                                "type": "number",
                                                "nullable": true
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": ["active", "completed", "overdue"]
                                            },
                                            "operations": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "string"
                                                        },
                                                        "goal_id": {
                                                            "type": "string"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "enum": ["deposit", "withdrawal"]
                                                        },
                                                        "amount": {
                                                            "type": "number"
                                                        },
                                                        "date": {
                                                            "type": "string",
                                                            "format": "date"
                                                        },
                                                        "note": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "required": [
                                                        "id",
                                                        "goal_id",
                                                        "type",
                                                        "amount",
                                                        "date",
                                                        "note",
                                                        "created_at",
                                                        "updated_at"
                                                    ]
                                                }
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "required": [
                                            "id",
                                            "workspace_id",
                                            "name",
                                            "target_amount",
                                            "initial_amount",
                                            "currency",
                                            "target_date",
                                            "plan_start_month",
                                            "archived_at",
                                            "current_amount",
                                            "remaining_amount",
                                            "progress_percentage",
                                            "remaining_months",
                                            "recommended_monthly_amount",
                                            "status",
                                            "created_at",
                                            "updated_at"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["Goals"]
            },
            "post": {
                "operationId": "GoalsController_create",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGoalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "workspace_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "target_amount": {
                                            "type": "number"
                                        },
                                        "initial_amount": {
                                            "type": "number"
                                        },
                                        "currency": {
                                            "type": "string"
                                        },
                                        "target_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "plan_start_month": {
                                            "type": "string"
                                        },
                                        "archived_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "current_amount": {
                                            "type": "number"
                                        },
                                        "remaining_amount": {
                                            "type": "number"
                                        },
                                        "progress_percentage": {
                                            "type": "number"
                                        },
                                        "remaining_months": {
                                            "type": "integer"
                                        },
                                        "recommended_monthly_amount": {
                                            "type": "number",
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": ["active", "completed", "overdue"]
                                        },
                                        "operations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "goal_id": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": ["deposit", "withdrawal"]
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "note": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "goal_id",
                                                    "type",
                                                    "amount",
                                                    "date",
                                                    "note",
                                                    "created_at",
                                                    "updated_at"
                                                ]
                                            }
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "workspace_id",
                                        "name",
                                        "target_amount",
                                        "initial_amount",
                                        "currency",
                                        "target_date",
                                        "plan_start_month",
                                        "archived_at",
                                        "current_amount",
                                        "remaining_amount",
                                        "progress_percentage",
                                        "remaining_months",
                                        "recommended_monthly_amount",
                                        "status",
                                        "created_at",
                                        "updated_at"
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Goals"]
            }
        },
...
        "/workspaces/{workspaceId}/goals/{goalId}/archive": {
            "post": {
                "operationId": "GoalsController_archive",
                "parameters": [
                    {
                        "name": "goalId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "workspace_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "target_amount": {
                                            "type": "number"
                                        },
                                        "initial_amount": {
                                            "type": "number"
                                        },
                                        "currency": {
                                            "type": "string"
                                        },
                                        "target_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "plan_start_month": {
                                            "type": "string"
                                        },
                                        "archived_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "current_amount": {
                                            "type": "number"
                                        },
                                        "remaining_amount": {
                                            "type": "number"
                                        },
                                        "progress_percentage": {
                                            "type": "number"
                                        },
                                        "remaining_months": {
                                            "type": "integer"
                                        },
                                        "recommended_monthly_amount": {
                                            "type": "number",
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": ["active", "completed", "overdue"]
                                        },
                                        "operations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "goal_id": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": ["deposit", "withdrawal"]
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "note": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "goal_id",
                                                    "type",
                                                    "amount",
                                                    "date",
                                                    "note",
                                                    "created_at",
                                                    "updated_at"
                                                ]
                                            }
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "workspace_id",
                                        "name",
                                        "target_amount",
                                        "initial_amount",
                                        "currency",
                                        "target_date",
                                        "plan_start_month",
                                        "archived_at",
                                        "current_amount",
                                        "remaining_amount",
                                        "progress_percentage",
                                        "remaining_months",
                                        "recommended_monthly_amount",
                                        "status",
                                        "created_at",
                                        "updated_at"
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Goals"]
            }
        },
...
    },
    "info": {
        "title": "Monqom API",
        "description": "The Monqom API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "CreateGoalDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Holiday",
                        "maxLength": 80
                    },
                    "target_amount": {
                        "type": "number",
                        "example": 12000,
                        "minimum": 0.01,
                        "maximum": 21474836.47
                    },
                    "initial_amount": {
                        "type": "number",
                        "example": 1500,
                        "minimum": 0,
                        "maximum": 21474836.47,
                        "default": 0
                    },
                    "target_date": {
                        "type": "string",
                        "example": "2027-08-19",
                        "format": "date"
                    },
                    "include_current_month": {
                        "type": "boolean",
                        "example": false,
                        "default": false
                    }
                },
                "required": ["name", "target_amount", "target_date", "include_current_month"]
            },
...
        }
    }
}